refactor(auth): 인증 훅과 마이페이지 에러 세션 의존성 정리
